+++ examples/resources/pingfederate_sp_authentication_policy_contract_mappings/resource.tf
@@ -1,4 +1,3 @@
-
 resource "pingfederate_authentication_policy_contract" "authenticationPolicyContractExample" {
   core_attributes     = [{ name = "subject" }]
   extended_attributes = [{ name = "extended_attribute" }, { name = "extended_attribute2" }]